% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/unitlabels.R
\name{uls}
\alias{uls}
\title{Insert parameter and unit labels in different markup languges}
\usage{
uls(sh, output = "R")
}
\arguments{
\item{sh}{User defined 'short-hand' code for a parameter, unit or 'parameter.unit' combination}

\item{output}{The desired markup languages to be returned}
}
\value{
A character string in the desired markup representing the requested label. \code{uls} returns the quantity symbol, \code{ulu} returns the quantity units, \code{ul} returns the combined quantity symbol and unit, \code{ulln} returns the quantity long name and \code{ulsn} returns the quantity standard name
}
\description{
Given a user defined 'short-hand' code for a parameter-unit this functions returns a label in the selected markup language.
}
\details{
User defined 'short-hand' code for a parameter-unit combinations and their conversions into different markup languages are stored in a data file \code{data/definitions.csv}. Edit this file to add or adjust definitions. Short hand code can take any form, but may not contain R's reserved words (\url{http://cran.r-project.org/doc/manuals/r-release/R-lang.html#Reserved-words})
}
\examples{
uls(sh="T", output="R")
uls(sh="T", output="md")
ulu("degC", output="R")
ulu("degC", output="md")
ul("T.degC", output="R")
ul("T.degC", output="md")
}

